requirements.txt
openai-whisper
yt-dlp
pyannote.audio==2.1.1
pydub
ffmpeg-python
speechbrain
openai
faster-whisper == 0.10.1
soundfile